refactor: Move guards to common/
Same mentality as below, these are used all over the place and don't need to be in auth/
