+++ server/src/modules/auth/auth.controller.ts
@@ -4,7 +4,7 @@ import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import * as passport from 'passport';
 import { Public } from '../../common/decorators/public.decorator';
 import { AuthService } from './auth.service';
-import { SteamWebAuthGuard } from './guard/steam-web-auth.guard';
+import { SteamWebAuthGuard } from '../../common/guards/steam-web-auth.guard';
 import { User } from '@prisma/client';
 import { SteamAuthService } from './steam-auth.service';
 import { JWTResponseDto } from '../../common/dto/jwt-response.dto';
