+++ server/src/modules/maps/maps.module.ts
@@ -5,7 +5,7 @@ import { RepoModule } from '../repo/repo.module';
 import { AuthModule } from '../auth/auth.module';
 import { FileStoreModule } from '../filestore/file-store.module';
 import { APP_GUARD } from '@nestjs/core';
-import { RolesGuard } from '../auth/guard/roles.guard';
+import { RolesGuard } from '../../common/guards/roles.guard';
 import { SessionModule } from '../session/session.module';
 import { XpSystemsModule } from '../xp-systems/xp-systems.module';
